feat: add checkout page
