+++ app/src/main/java/id/my/miruza/locavest/CartItemAdapter.kt
@@ -1,20 +1,12 @@
 package id.my.miruza.locavest
 
-import android.content.Context
-import android.os.Parcel
-import android.os.Parcelable
 import android.util.Log
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import id.my.miruza.locavest.databinding.LayoutCartItemBinding
-import kotlinx.parcelize.Parceler
-import kotlinx.parcelize.Parcelize
 
 class CartItemAdapter(
     private val items: List<CartItem>
@@ -56,34 +48,34 @@ class CartItemAdapter(
     override fun getItemCount(): Int = items.size
 }
 
-@Parcelize
-data class CartItem(
-    var image: String,
-    var name: String,
-    var pricePer: Float,
-    var priceUnit: String,
-    var amount: Int,
-): Parcelable {
-    constructor(parcel: Parcel) : this(
-        parcel.readString()!!,
-        parcel.readString()!!,
-        parcel.readFloat(),
-        parcel.readString()!!,
-        parcel.readInt()
-    )
-
-    companion object : Parceler<CartItem> {
-
-        override fun CartItem.write(parcel: Parcel, flags: Int) {
-            parcel.writeString(image)
-            parcel.writeString(name)
-            parcel.writeFloat(pricePer)
-            parcel.writeString(priceUnit)
-            parcel.writeInt(amount)
-        }
-
-        override fun create(parcel: Parcel): CartItem {
-            return CartItem(parcel)
-        }
-    }
-}
+//@Parcelize
+//data class CartItem(
+//    var image: String,
+//    var name: String,
+//    var pricePer: Float,
+//    var priceUnit: String,
+//    var amount: Int,
+//): Parcelable {
+//    constructor(parcel: Parcel) : this(
+//        parcel.readString()!!,
+//        parcel.readString()!!,
+//        parcel.readFloat(),
+//        parcel.readString()!!,
+//        parcel.readInt()
+//    )
+//
+//    companion object : Parceler<CartItem> {
+//
+//        override fun CartItem.write(parcel: Parcel, flags: Int) {
+//            parcel.writeString(image)
+//            parcel.writeString(name)
+//            parcel.writeFloat(pricePer)
+//            parcel.writeString(priceUnit)
+//            parcel.writeInt(amount)
+//        }
+//
+//        override fun create(parcel: Parcel): CartItem {
+//            return CartItem(parcel)
+//        }
+//    }
+//}
