+++ app/src/main/res/values/strings.xml
@@ -6,4 +6,6 @@
     <string name="btn_launch_checkout_text">Launch Checkout Screen</string>
     <string name="cart_item_pricing_format">%.2f Rp / %s</string>
     <string name="btn_checkout_text">Checkout</string>
+    <string name="text_placeholder">Lorem Ipsum</string>
+    <string name="string_total_price_format">%.2f IDR</string>
 </resources>
